\name{cartodb.test}
\alias{cartodb.test}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Tests that cartodb connection is valid
}
\description{
Tests that cartodb connection is valid. Will also test secure connection if an API key is supplied
}
\usage{
cartodb.test()
}
\examples{
cartodb.test()
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
