\name{cartodb.tiles.base}
\alias{cartodb.tiles.base}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Turns the user supplied options into the base urls needed for Maps API calls.
}
\description{
A set of predefined ways of turning cartodb.R options into actionable URLs.
}
\usage{
cartodb.tiles.base()
}
\author{
Andrew Hill <andrew@vizzuality.com>
}
\examples{
cartodb.tiles.base()
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
