\name{jsonToDataFrame}
\alias{jsonToDataFrame}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
jsonToDataFrame
}
\description{
Turns a JSON string into a data matrix. Based off of work 
}
\usage{
jsonToDataFrame(json)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{json}{
  JSON string
}=
}
\examples{
jsonToDataFrame("{g: 1}")
}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
